Implementation that actually looks at the target version we set
