+++ crates/red_knot_module_resolver/src/lib.rs
@@ -3,6 +3,7 @@ mod module;
 mod module_name;
 mod path;
 mod resolver;
+mod supported_py_version;
 mod typeshed;
 
 pub use db::{Db, Jar};
@@ -13,4 +14,5 @@ pub use path::{
     SitePackagesPathBuf, StandardLibraryPath, StandardLibraryPathBuf,
 };
 pub use resolver::{resolve_module, set_module_resolution_settings, ModuleResolutionSettings};
-pub use typeshed::{SupportedPyVersion, TypeshedVersions};
+pub use supported_py_version::SupportedPyVersion;
+pub use typeshed::TypeshedVersions;
