+++ crates/red_knot_module_resolver/src/typeshed.rs
@@ -1,9 +1,7 @@
-mod supported_py_version;
 mod versions;
 
-pub(crate) use versions::TypeshedVersionsQueryResult;
 pub use versions::TypeshedVersions;
-pub use supported_py_version::SupportedPyVersion;
+pub(crate) use versions::TypeshedVersionsQueryResult;
 
 #[cfg(test)]
 mod tests {
